+++ test/tortoise/connection/receiver_test.exs
@@ -13,10 +13,10 @@ defmodule Tortoise.Connection.ReceiverTest do
   def setup_receiver(context) do
     opts = [client_id: context.client_id]
     {:ok, client_socket, server_socket} = Tortoise.Integration.TestTCPTunnel.new()
-    {:ok, _pid} = Receiver.start_link(opts)
+    {:ok, receiver_pid} = Receiver.start_link(opts)
     {:ok, _pid} = Transmitter.start_link(opts)
     :ok = Receiver.handle_socket(context.client_id, {Tortoise.Transport.Tcp, client_socket})
-    {:ok, %{client: client_socket, server: server_socket}}
+    {:ok, %{receiver_pid: receiver_pid, client: client_socket, server: server_socket}}
   end
 
   def setup_controller(context) do
@@ -52,6 +52,32 @@ defmodule Tortoise.Connection.ReceiverTest do
 
       assert_receive {:"$gen_cast", {:incoming, data}}, 10000
       assert ^package = Package.decode(data)
+    end
+
+    test "very slow connection", context do
+      Process.flag(:trap_exit, true)
+      receiver_pid = context.receiver_pid
+      # Send a byte, one at a time, into the receiver. This has been
+      # observed in the wild and it caused a faulty pattern match to
+      # report it as an invalid_header_length protocol violation
+      :ok = :gen_tcp.send(context.server, <<0b11010000>>)
+      refute_receive {:EXIT, ^receiver_pid, {:protocol_violation, :invalid_header_length}}, 400
+      :ok = :gen_tcp.send(context.server, <<0>>)
+      assert_receive {:"$gen_cast", {:incoming, data}}, 10000
+      assert %Package.Pingresp{} = Package.decode(data)
+    end
+  end
+
+  describe "invalid packages" do
+    setup [:setup_receiver]
+
+    test "invalid header length", context do
+      Process.flag(:trap_exit, true)
+      receiver_pid = context.receiver_pid
+      # send too many bytes into the receiver, the header parser
+      # should throw a protocol violation on this
+      :ok = :gen_tcp.send(context.server, <<1, 255, 255, 255, 255, 0>>)
+      assert_receive {:EXIT, ^receiver_pid, {:protocol_violation, :invalid_header_length}}
     end
   end
 end